Implement Zeroconf-based network discovery
Add a a dummy client-server pair. These use HTTP/2 for simplicity.
The server is very dumb and will probably be totally replaced.

The client comes with a decent implementation of sending jobs from a
queue to multiple servers while more servers dynamically join. It can
be the basis of a more complete implementation.

The version of Zeroconf being used here is Bonjour, but it's not
Apple-specific (it's a complete reimplementation; even works on Windows).

The Zeroconf code is deliberately kept separate from the client/server.