Implement zero-config network discovery and mock client/daemon pair (#34)
* Add Bonjour

* Implement Zeroconf-based network discovery

Add a a dummy client-server pair. These use HTTP/2 for simplicity.
The server is very dumb and will probably be totally replaced.

The client comes with a decent implementation of sending jobs from a
queue to multiple servers while more servers dynamically join. It can
be the basis of a more complete implementation.

The version of Zeroconf being used here is Bonjour, but it's not
Apple-specific (it's a complete reimplementation; even works on Windows).

The Zeroconf code is deliberately kept separate from the client/server.

* Make ESLint aware of TSDoc, Jest, and Prettier

* Separate binaries from library code

Exposes executables as commandline programs: junknetd and junknet-make
Exposes the module as a JS library

* Consistent filename casing; add constants to lib

* Build good-looking doc website from TSDoc comments

* More thorough documentation

* Stupid Windows case insensitive filenames

* Update typedoc (and other packages)

* Resolve review comments

* Exclude Vim swap files