Fix: `--help`'s exit code depends on undefined behavior.
We should check for `--help` being passed *before* we check for invalid
arguments, otherwise nobody would be able to see the help message
without already knowing how to use Junknet. The previous implementation
depended on undefined behavior in the `CommandLineInterface` module.

Also, I've changed the behavior of the "invalid arguments" case. We:

1. Print a human-readable error message (*not* a stack trace) to stderr.
2. Exit the process with an error status code.